+++ www/src/components/docs/page-with-sidebar.js
@@ -2,8 +2,8 @@
 import { jsx } from "theme-ui"
 import { Fragment } from "react"
 
-import { getItemList } from "../utils/sidebar/item-list"
-import StickyResponsiveSidebar from "./docs/sidebar/sticky-responsive-sidebar"
+import { getItemList } from "../../utils/sidebar/item-list"
+import StickyResponsiveSidebar from "./sidebar/sticky-responsive-sidebar"
 
 export default ({ children, enableScrollSync, location }) => {
   const itemList = getItemList(location.pathname)
