+++ www/src/components/homepage/homepage-newsletter.js
@@ -2,7 +2,7 @@ import React from "react"
 import styled from "@emotion/styled"
 
 import HomepageSection from "./homepage-section"
-import EmailCaptureForm from "../../components/email-capture-form"
+import EmailCaptureForm from "../../components/mdx/email-capture-form"
 
 import { NewsletterFormOrnament } from "../../assets/ornaments"
