+++ www/src/components/layout/banner.js
