+++ www/src/components/layout/index.js
@@ -4,15 +4,15 @@ import React from "react"
 
 import { Global } from "@emotion/core"
 
-import { globalStyles } from "../utils/styles/global"
-import { breakpointGutter } from "../utils/styles"
-import Banner from "../components/banner"
-import Navigation from "../components/navigation"
-import MobileNavigation from "../components/navigation-mobile"
-import SiteMetadata from "../components/site-metadata"
-import SkipNavLink from "../components/skip-nav-link"
-import "../assets/fonts/futura"
-import { defaultLang } from "../utils/i18n"
+import { globalStyles } from "../../utils/styles/global"
+import { breakpointGutter } from "../../utils/styles"
+import Banner from "./banner"
+import Navigation from "./navigation"
+import MobileNavigation from "./navigation-mobile"
+import SiteMetadata from "./site-metadata"
+import SkipNavLink from "./skip-nav-link"
+import "../../assets/fonts/futura"
+import { defaultLang } from "../../utils/i18n"
 
 export const LocaleContext = React.createContext(defaultLang)
