+++ www/src/components/layout/navigation-mobile.js
@@ -8,9 +8,9 @@ import {
   TutorialIcon,
   PluginsIcon,
   ShowcaseIcon,
-} from "../assets/icons"
+} from "../../assets/icons"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
-import { svgStyles } from "../utils/styles"
+import { svgStyles } from "../../utils/styles"
 
 const getProps = ({ isPartiallyCurrent }) => {
   return {
