+++ www/src/components/layout/navigation.js
@@ -4,14 +4,14 @@ import { useColorMode } from "theme-ui"
 import GithubIcon from "react-icons/lib/go/mark-github"
 import TwitterIcon from "react-icons/lib/fa/twitter"
 
-import Link from "../components/localized-link"
+import Link from "../localized-link"
 import SearchForm from "./search-form"
-import DiscordIcon from "../components/discord"
-import logo from "../assets/logo.svg"
-import logoInverted from "../assets/logo-inverted.svg"
+import DiscordIcon from "./discord"
+import logo from "../../assets/logo.svg"
+import logoInverted from "../../assets/logo-inverted.svg"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
-import { breakpointGutter } from "../utils/styles"
-import DarkModeToggle from "../components/dark-mode-toggle"
+import { breakpointGutter } from "../../utils/styles"
+import DarkModeToggle from "../dark-mode-toggle"
 
 // what we need to nudge down the navItems to sit
 // on the baseline of the logo's wordmark
