+++ www/src/components/layout/skip-nav-link.js
@@ -2,7 +2,7 @@
 import { jsx } from "theme-ui"
 import { SkipNavLink } from "@reach/skip-nav"
 
-import { visuallyHidden } from "../utils/styles"
+import { visuallyHidden } from "../../utils/styles"
 
 const Link = () => (
   <SkipNavLink
