+++ www/src/components/mdx/hubspot-form.js
@@ -2,7 +2,7 @@
 import { jsx } from "theme-ui"
 import { Component } from "react"
 
-import HubspotForm from "../react-hubspot-form"
+import HubspotForm from "./react-hubspot-form"
 import { formInput, formInputFocus, buttonStyles } from "../../utils/styles"
 
 export default class GatsbyHubspotForm extends Component {
