--- www/src/components/search-form/index.js
@@ -1,1 +1,0 @@
-export { default } from "./search-form"
