+++ www/src/gatsby-plugin-theme-ui/components.js
@@ -2,11 +2,11 @@ import React from "react"
 
 import GuideList from "../components/mdx/guide-list.js"
 import HubspotForm from "../components/mdx/hubspot-form"
-import Pullquote from "../components/shared/pullquote"
-import EggheadEmbed from "../components/shared/egghead-embed"
+import Pullquote from "../components/mdx/pullquote"
+import EggheadEmbed from "../components/mdx/egghead-embed"
 import DateChart from "../components/mdx/chart"
 import LayerModel from "../components/mdx/layer-model"
-import EmailCaptureForm from "../components/email-capture-form"
+import EmailCaptureForm from "../components/mdx/email-capture-form"
 import HorizontalNavList from "../components/mdx/horizontal-nav-list"
 import CodeBlock from "../components/code-block"
 import MdxLink from "../components/mdx/mdx-link"
