+++ www/src/pages/404.js
@@ -2,7 +2,7 @@ import React from "react"
 import Container from "../components/container"
 import Layout from "../components/layout"
 import { Link } from "gatsby"
-import FooterLinks from "../components/shared/footer-links"
+import FooterLinks from "../components/layout/footer-links"
 
 class FourOhFour extends React.Component {
   render() {
