+++ www/src/pages/blog/tags.js
@@ -12,10 +12,14 @@ import Button from "../../components/button"
 import Layout from "../../components/layout"
 import Container from "../../components/container"
 import SearchIcon from "../../components/search-icon"
-import FooterLinks from "../../components/shared/footer-links"
+import FooterLinks from "../../components/layout/footer-links"
 import { TAGS_AND_DOCS } from "../../data/tags-docs"
 import { themedInput } from "../../utils/styles"
-import { colors, space, mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
+import {
+  colors,
+  space,
+  mediaQueries,
+} from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 
 const POPULAR_TAGS = [
   `themes`,
