+++ www/src/pages/features/cms.js
@@ -4,11 +4,11 @@ import { graphql } from "gatsby"
 
 import Button from "../../components/button"
 import Layout from "../../components/layout"
-import PageWithSidebar from "../../components/page-with-sidebar"
+import PageWithSidebar from "../../components/docs/page-with-sidebar"
 import EvaluationTable from "../../components/features/evaluation-table"
 import { getFeaturesData } from "../../utils/get-csv-features-data"
 import Container from "../../components/container"
-import FooterLinks from "../../components/shared/footer-links"
+import FooterLinks from "../../components/layout/footer-links"
 import FeaturesFooter from "../../components/features/features-footer"
 import LegendTable from "../../components/features/legend-table"
 import CompareButton from "../../components/features/compare-button"
