+++ www/src/pages/features/jamstack.js
@@ -7,14 +7,14 @@ import Layout from "../../components/layout"
 import EvaluationTable from "../../components/features/evaluation-table"
 import { getFeaturesData } from "../../utils/get-csv-features-data"
 import Container from "../../components/container"
-import FooterLinks from "../../components/shared/footer-links"
+import FooterLinks from "../../components/layout/footer-links"
 import FeaturesFooter from "../../components/features/features-footer"
 import LegendTable from "../../components/features/legend-table"
 import CompareButton from "../../components/features/compare-button"
 import Breadcrumb from "../../components/docs/docs-breadcrumb"
 import featureComparisonOptions from "../../data/features/comparison-options.json"
 import useComparisonState from "../../hooks/use-comparison-state"
-import PageWithSidebar from "../../components/page-with-sidebar"
+import PageWithSidebar from "../../components/docs/page-with-sidebar"
 
 const FeaturesHeader = () => (
   <section>
