+++ www/src/templates/template-api-markdown.js
@@ -10,10 +10,10 @@ import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { sortBy } from "lodash-es"
 
 import Layout from "../components/layout"
-import PageWithSidebar from "../components/page-with-sidebar"
-import MarkdownPageFooter from "../components/markdown-page-footer"
+import PageWithSidebar from "../components/docs/page-with-sidebar"
+import MarkdownPageFooter from "../components/docs/markdown-page-footer"
 import DocSearchContent from "../components/docsearch-content"
-import FooterLinks from "../components/shared/footer-links"
+import FooterLinks from "../components/layout/footer-links"
 import Breadcrumb from "../components/docs/docs-breadcrumb"
 import Container from "../components/container"
 import PrevAndNext from "../components/prev-and-next"
