+++ www/src/templates/template-blog-list.js
@@ -10,8 +10,8 @@ import Button from "../components/button"
 import Container from "../components/container"
 import BlogPostPreviewItem from "../components/blog/blog-post-preview-item"
 import Pagination from "../components/blog/pagination"
-import EmailCaptureForm from "../components/email-capture-form"
-import FooterLinks from "../components/shared/footer-links"
+import EmailCaptureForm from "../components/mdx/email-capture-form"
+import FooterLinks from "../components/layout/footer-links"
 
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 import { pullIntoGutter, breakpointGutter } from "../utils/styles"
