+++ www/src/templates/template-docs-markdown.js
@@ -7,11 +7,11 @@ import { MDXRenderer } from "gatsby-plugin-mdx"
 import { mediaQueries } from "gatsby-design-tokens/dist/theme-gatsbyjs-org"
 
 import Layout from "../components/layout"
-import PageWithSidebar from "../components/page-with-sidebar"
-import MarkdownPageFooter from "../components/markdown-page-footer"
+import PageWithSidebar from "../components/docs/page-with-sidebar"
+import MarkdownPageFooter from "../components/docs/markdown-page-footer"
 import DocSearchContent from "../components/docsearch-content"
 import TableOfContents from "../components/docs/docs-table-of-contents"
-import FooterLinks from "../components/shared/footer-links"
+import FooterLinks from "../components/layout/footer-links"
 import Breadcrumb from "../components/docs/docs-breadcrumb"
 import Container from "../components/container"
 import PrevAndNext from "../components/prev-and-next"
