+++ www/src/templates/template-feature-comparison.js
@@ -4,8 +4,8 @@ import { Component } from "react"
 import { Helmet } from "react-helmet"
 import { css } from "@emotion/core"
 import Layout from "../components/layout"
-import PageWithSidebar from "../components/page-with-sidebar"
-import FooterLinks from "../components/shared/footer-links"
+import PageWithSidebar from "../components/docs/page-with-sidebar"
+import FooterLinks from "../components/layout/footer-links"
 import Container from "../components/container"
 import EvaluationTable from "../components/features/evaluation-table"
 import LogoDictionary from "../components/features/logo-dictionary"
